+++ git-webui/src/share/git-webui/webui/js/git-webui.js
@@ -2588,13 +2588,19 @@ webui.NewChangedFilesView = function(workspaceView) {
                 });
                 $("#commitBtn").off("click");
                 $("#commitBtn").on("click", function() {
-                    if (selectedItemsFromOtherUser.length > 0) {
-                        self.confirmActionOnOtherUsersChanges("commit");
-                    } else {
-                        var commitMessage = $('#commitMsg').val();
-                        self.commit(commitMessage, $("#commitMsgDetail").val());
-                    }
-                    
+                    $.get("api/basic-and-default", function (data) {
+                        var basicAndDefault = JSON.parse(data)["basic-and-default"]
+                        if (basicAndDefault == "1") {
+                            self.noCommitsOnDefault();
+                        } else {
+                            if (selectedItemsFromOtherUser.length > 0) {
+                                self.confirmActionOnOtherUsersChanges("commit");
+                            } else {
+                                var commitMessage = $('#commitMsg').val();
+                                self.commit(commitMessage, $("#commitMsgDetail").val());
+                            }
+                        }
+                    })                    
                 });
 
                 $("#amendBtn").off("click");
@@ -2738,6 +2744,54 @@ webui.NewChangedFilesView = function(workspaceView) {
             removePopup(popup);
         });
     }
+
+    self.noCommitsOnDefault = function () {
+        function removePopup(popup) {
+            $(popup).children(".modal-fade").modal("hide");
+            $(".modal-backdrop").remove();
+            $("#noCommitsDefault").remove();
+        }
+
+        var popup = $(
+            '<div class="modal fade" tabindex="-1" id="noCommitsDefault" role="dialog" data-backdrop="static">' +
+                '<div class="modal-dialog modal-md" role="document">' +
+                    '<div class="modal-content">' + 
+                        '<div class="modal-header">' +
+                            '<h5 class="modal-title">Cannot commit to Default Branch</h5>' +
+                            '<button type="button" class="btn btn-default close" data-dismiss="modal">' + webui.largeXIcon + '</button>' +
+                        '</div>' +
+                        '<div class="modal-body">' + 
+                            '<div class="row">' +
+                                '<div class="col-sm-1">' +
+                                    webui.warningIcon +
+                                '</div>' +
+                                '<div class="col-sm-11">' +
+                                    '<p>You cannot commit to the default merge branch while using basic mode. Please switch to another branch.</p>' + 
+                                '</div>' +
+                            '</div>' +
+                        '</div>' +
+                        '<div class="modal-footer"></div>' +
+                    '</div>' + 
+                '</div>' +
+            '</div>'
+        )[0];
+
+        $("body").append(popup);
+
+        var popupFooter = $(".modal-footer", popup)[0];
+        webui.detachChildren(popupFooter);
+
+        $(
+            '<button class="btn btn-sm btn-secondary action-btn" id="noCommitDefaultButton">Ok</button>'
+        ).appendTo(popupFooter);
+
+        $(popup).modal('show');
+
+        $("#noCommitsDefault").find(".close, #noCommitDefaultButton").click(function() {
+            removePopup(popup);
+        })
+
+    };
 
     self.confirmActionOnOtherUsersChanges = function(action) {
             function removeWarningModal(popup) {