+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed display of other users' username in workspace view on Unix (#530)
+- Prevent direct commits to default merge branch in basic mode (#484)
 
 ## [2.6.0] - 2024-10-07